Display data for search city and refractor code for reusability

+++ weather_app/lib/model/weather_model.dart
@@ -1,5 +1,5 @@
 class WeatherModel {
-  final int temp;
+  final double temp;
   final String weatherDesc;
   final double windSpeed;
   final String weatherMain;
